+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for firebase auth service
 - Added support for firebase storage service
 - Rework of the API in line with flutterfire packages 
+- `.info` location with `connected`, `authenticated` and `serverTimeOffset`
 
 ## 0.7.15